refactor: 拆分被动群聊捕获逻辑
